Fix "out of buffers" crash (#3783)
Remove BufferManager implementation. Use MemoryManager allocation instead and re-use buffers where they are allocated (AudioPort.cpp & PlayHandle.cpp)

+++ include/MemoryHelper.h
@@ -31,7 +31,7 @@
 class MemoryHelper {
 public:
 
-	static void* alignedMalloc( int );
+	static void* alignedMalloc( size_t );
 
 	static void alignedFree( void* );
